+++ src/tests/Test.App/Program.cs
@@ -1,8 +1,30 @@
+using EntityGraphQL;
 using EntityGraphQL.AspNet;
 
 var builder = WebApplication.CreateBuilder(args);
 
-builder.Services.AddGraphQLSchema<TestQueryType>();
+builder.Services.AddGraphQLValidator();
+builder.Services.AddGraphQLSchema<TestQueryType>(configure =>
+{
+    configure.ConfigureSchema = schema =>
+    {
+        schema
+            .Mutation()
+            .Add(
+                "mutationFail",
+                (TestQueryType db, IGraphQLValidator validator) =>
+                {
+                    validator.AddError("This is a test error");
+                    if (validator.HasErrors)
+                    {
+                        return null;
+                    }
+                    return db.Hello;
+                }
+            )
+            .IsNullable(false);
+    };
+});
 builder.Services.AddScoped<TestQueryType>();
 
 var app = builder.Build();